backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.35
pymysql==1.1.1

pandas==2.2.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1

ftfy==6.2.0
bleach==6.1.0
langdetect==1.0.9
requests==2.32.3
python-dotenv==1.0.1

